Move signin and signup to (auth)
