+++ src/app/[locale]/(auth)/action.ts
@@ -1,9 +1,24 @@
 'use server';
 
+import { createClient } from '@/lib/supabase/server';
 import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
 
-import { createClient } from '@/lib/supabase/server';
+export async function emailSignIn(email: string, password: string) {
+  const supabase = createClient();
+
+  const { error } = await supabase.auth.signInWithPassword({
+    email,
+    password,
+  });
+
+  if (error) {
+    redirect('/error');
+  }
+
+  revalidatePath('/', 'layout');
+  redirect('/');
+}
 
 export async function emailSignUp(email: string, password: string) {
   const supabase = createClient();
