+++ src/app/[locale]/(auth)/signin/components/UserSignInForm.tsx
@@ -8,7 +8,7 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { cn } from '@/lib/utils';
 
-import { emailSignIn } from '../action';
+import { emailSignIn } from '../../action';
 
 type UserAuthFormProps = Record<string, unknown> & React.HTMLAttributes<HTMLDivElement>;
