+++ src/app/[locale]/(auth)/signin/page.tsx
@@ -1,23 +1,22 @@
-import type { Metadata } from 'next';
-import Link from 'next/link';
-
-import { UserSignupForm } from '@/app/[locale]/signup/components/UserSignUpForm';
+import { UserSignInForm } from '@/app/[locale]/(auth)/signin/components/UserSignInForm';
 import { buttonVariants } from '@/components/ui/button';
 import { cn } from '@/lib/utils';
+import type { Metadata } from 'next';
+import Link from 'next/link';
 
 export const metadata: Metadata = {
   title: 'Authentication',
   description: 'Authentication forms built using the components.',
 };
 
-export default function SignUpPage({
+export default function SignInPage({
   params: { lang },
 }: Readonly<{
   params: { lang: string };
 }>) {
   return (
     <>
-      <div className="relative min-h-screen flex-col items-center justify-center grid">
+      <div className="relative hidden min-h-screen flex-col items-center justify-center md:grid">
         <Link
           href="/examples/authentication"
           className={cn(buttonVariants({ variant: 'ghost' }), 'absolute right-4 top-4 md:right-8 md:top-8')}
@@ -29,7 +28,7 @@ export default function SignUpPage({
             <h1 className="text-2xl font-semibold tracking-tight">Create an account</h1>
             <p className="text-sm text-muted-foreground">Enter your email below to create your account</p>
           </div>
-          <UserSignupForm />
+          <UserSignInForm />
           <p className="px-8 text-center text-sm text-muted-foreground">
             By clicking continue, you agree to our{' '}
             <Link href="/terms" className="underline underline-offset-4 hover:text-primary">
