+++ src/app/[locale]/(auth)/signup/components/UserSignUpForm.tsx
@@ -7,8 +7,7 @@ import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { cn } from '@/lib/utils';
-
-import { discordSignIn, emailSignUp } from '../action';
+import { discordSignIn, emailSignUp } from '../../action';
 
 type UserAuthFormProps = Record<string, unknown> & React.HTMLAttributes<HTMLDivElement>;
