+++ src/app/[locale]/(auth)/signup/page.tsx
@@ -1,23 +1,22 @@
-import type { Metadata } from 'next';
-import Link from 'next/link';
-
-import { UserSignInForm } from '@/app/[locale]/signin/components/UserSignInForm';
 import { buttonVariants } from '@/components/ui/button';
 import { cn } from '@/lib/utils';
+import type { Metadata } from 'next';
+import Link from 'next/link';
+import { UserSignupForm } from './components/UserSignUpForm';
 
 export const metadata: Metadata = {
   title: 'Authentication',
   description: 'Authentication forms built using the components.',
 };
 
-export default function SignInPage({
+export default function SignUpPage({
   params: { lang },
 }: Readonly<{
   params: { lang: string };
 }>) {
   return (
     <>
-      <div className="relative hidden min-h-screen flex-col items-center justify-center md:grid">
+      <div className="relative min-h-screen flex-col items-center justify-center grid">
         <Link
           href="/examples/authentication"
           className={cn(buttonVariants({ variant: 'ghost' }), 'absolute right-4 top-4 md:right-8 md:top-8')}
@@ -29,7 +28,7 @@ export default function SignInPage({
             <h1 className="text-2xl font-semibold tracking-tight">Create an account</h1>
             <p className="text-sm text-muted-foreground">Enter your email below to create your account</p>
           </div>
-          <UserSignInForm />
+          <UserSignupForm />
           <p className="px-8 text-center text-sm text-muted-foreground">
             By clicking continue, you agree to our{' '}
             <Link href="/terms" className="underline underline-offset-4 hover:text-primary">
