--- src/app/[locale]/signin/action.ts
@@ -1,22 +1,0 @@
-'use server';
-
-import { revalidatePath } from 'next/cache';
-import { redirect } from 'next/navigation';
-
-import { createClient } from '@/lib/supabase/server';
-
-export async function emailSignIn(email: string, password: string) {
-  const supabase = createClient();
-
-  const { error } = await supabase.auth.signInWithPassword({
-    email,
-    password,
-  });
-
-  if (error) {
-    redirect('/error');
-  }
-
-  revalidatePath('/', 'layout');
-  redirect('/');
-}
